improve/fix widevine initial start
There it becomes it a crash if within ~30-32 seconds after start and dlopen of widevine
Kodi does exit.

If now a load was done becomes a delay added if Kodi becomes stopped.
As improve is when, via inputstream.adaptive the widewine is present
becomes the own manifest creation inhibited and the dlopen not needed.